EduTrack/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.4
debugpy==1.4.1
decorator==5.0.9
distlib==0.3.8
distro==1.9.0
Django==4.2.13
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
filelock==3.15.1
frozenlist==1.4.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
idna==3.7
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
matplotlib-inline==0.1.2
multidict==6.0.5
openai==0.28.0
parso==0.8.2
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.2.2
prompt-toolkit==3.0.19
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.9.0
python-dateutil==2.8.2
pywin32==301
pyzmq==22.1.0
requests==2.32.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.0
tornado==6.1
tqdm==4.66.4
traitlets==5.0.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.2
wcwidth==0.2.5
webencodings==0.5.1
yarl==1.9.4
